+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ Added
 * Raise ValueError exception if longitude is outside [-180, 180] or latitude is outside [-90, 90] (#236, #218 @Andrey-Raspopov)
 * optional ``timeout`` attribute to avoid indefinite wait on response from the server (#256, @viktorbahr)
 * Parsing the ``Online``, ``CreationDate`` and ``IngestionDate`` fields of an OData response
+* Trying to download an offline product from the Copernicus Open Access Hub triggers its retrieval from the long term archive.
+  Downloading of the product is **not** scheduled.
 
 Changed
 ~~~~~~~